+++ tests/extend/AwareExtend.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the hyn/multi-tenant package.
+ *
+ * (c) Daniël Klabbers <daniel@klabbers.email>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @see https://laravel-tenancy.com
+ * @see https://github.com/hyn/multi-tenant
+ */
+
+namespace Hyn\Tenancy\Tests\Extend;
+
+use Hyn\Tenancy\Traits\TenantAwareConnection;
+
+class AwareExtend extends TenantExtend
+{
+    use TenantAwareConnection;
+}