+++ changelog.md
@@ -1,3 +1,6 @@
+- 5.2.5
+    - #524, added tenant aware logger.
+    - #534, added tenant aware connection trait, which makes Eloquent models use the tenant connection if it exists, otherwise the system connection.
 - 5.2.4
     - #400, fixed issue where instantiating a model again would not override its connection to system or tenant if configured.
 - 5.2.3